Move _get_tr from plot to bids.
